Avancement

+++ server/database/migrations/2017_06_10_195952_add_workedHoursParDay_users_table.php
@@ -15,7 +15,7 @@
         {
             Schema::table('users', function (Blueprint $table)
             {
-                $table->integer('worked_hours_per_day')->default(7)->after('remember_token');
+                $table->decimal('worked_hours_per_day', 10, 2)->default(7)->after('remember_token');
             });
         }
