+++ server/database/migrations/2017_06_10_200701_add_workedDuration_taskUser_table.php
@@ -15,7 +15,7 @@
         {
             Schema::table('task_user', function (Blueprint $table)
             {
-                $table->integer('worked_duration')->default(0)->after('duration');
+                $table->decimal('worked_duration', 10, 2)->default(0)->after('duration');
             });
         }
